requirements.txt
streamlit==1.10.0
openai==0.27.0
pinecone-client==2.0.0
pandas==1.4.0
